+++ example/example.dart
@@ -5,3 +5,6 @@ var promptpayDataWithoutAmount = PromptPay.generateQRData("0812345678");
 
 final qrData = "00020101021129370016A000000677010111011300668123456785802TH53037645406123.4563043045";
 var accountNumber = PromptPay.getAccountNumberFromQRData(qrData);
+
+// add amount to existing qr code
+final newQRCode = PromptPay.getPromptPayQRWithNewAmount("00020101021129370016A000000677010111011300668123456785802TH53037645406123.4563043045", 200.04);